Added splash screen startup
