+++ src/Strem.Infrastructure/Services/Api/IApiWebHost.cs
@@ -2,6 +2,7 @@
 
 public interface IApiWebHost
 {
+    bool IsRunning { get; }
     void StartHost();
     void StopHost();
 }
