+++ src/Strem.UI/Pages/Flows.razor
@@ -1,4 +1,4 @@
-﻿@page "/"
+﻿@page "/flows"
 
 @using Strem.Core.Flows
 @using Strem.Infrastructure.Services.Persistence.Flows
