+++ src/Strem.UI/Pages/Test.razor
@@ -1,4 +1,4 @@
-﻿@page "/Test"
+﻿@page "/test"
 @using Strem.Core.Types
 
 <div class="container is-fullwidth">
