+++ src/Strem.UI/Shared/MainLayout.razor
@@ -1,14 +1,30 @@
 ﻿@inherits LayoutComponentBase
 
+@inject NavigationManager NavManager
+
 <section class="hero is-dark is-fullheight">
-  <div class="hero-head has-background-grey-dark">
-    <span class="tag is-dark strem-logo">Strem - v@(GetType().Assembly.GetName().Version) (Alpha)</span>
-    <div class="container">
-      <TopMenu/>
+  @if (!IsSplashScreen)
+  {
+    <div class="hero-head has-background-grey-dark">
+      <span class="tag is-dark strem-logo">Strem - v@(GetType().Assembly.GetName().Version) (Alpha)</span>
+      <div class="container">
+        <TopMenu/>
+      </div>
     </div>
-  </div>
+  }
+
   <div class="hero-body is-align-items-stretch">
     @Body
   </div>
-  <div class="hero-foot"></div>
+
+  @if (!IsSplashScreen)
+  {
+    <div class="hero-foot"></div>
+  }
 </section>
+
+@code {
+
+  public bool IsSplashScreen => (NavManager.Uri == NavManager.BaseUri);
+  
+}
