+++ src/Strem.UI/Shared/TopMenu.razor
@@ -1,7 +1,7 @@
 ﻿<div class="tabs is-centered is-boxed">
   <ul class="pt-4">
     <li>
-      <NavLink href="/" ActiveClass="has-background-info-light has-text-black-bis" Match="NavLinkMatch.All">
+      <NavLink href="flows" ActiveClass="has-background-info-light has-text-black-bis" Match="NavLinkMatch.All">
         <span class="icon is-small"><i class="fas fa-code-commit" aria-hidden="true"></i></span>
         <span>Flows</span>
       </NavLink>
