+++ src/Strem/Program.cs
@@ -1,7 +1,10 @@
+using System.Reactive.Concurrency;
+using System.Reactive.Linq;
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Photino.Blazor;
+using PhotinoNET;
 using Strem.Core.DI;
 using Strem.Core.Events;
 using Strem.Core.Events.Bus;
@@ -29,73 +32,21 @@ public class Program
         _ = typeof(TwitchModule).Assembly;
         _ = typeof(OBSModule).Assembly;
 
-        LoadExternalPlugins();
-    }
-
-    public static void LoadExternalPlugins()
-    {
-        var pluginsDirectory = "Plugins";
-        if (!Directory.Exists(pluginsDirectory)) { return; }
-        
-        var pluginFiles = Directory.GetFileSystemEntries(pluginsDirectory, "*.dll");
-        if(pluginFiles.Length == 0) { return; }
-
-        foreach (var pluginFile in pluginFiles)
-        {
-            try { Assembly.LoadFile(pluginFile); }
-            catch(Exception ex) { StartupErrors.Add($"Failed to load Plugin {pluginFile}: {ex.Message}"); }
-        }
-    }
-    
-    public static Type[] GetAllDependencyModules()
-    {
-        LoadAllPluginModules();
-        var assemblies = AppDomain.CurrentDomain.GetAssemblies();
-        
-        return assemblies
-            .SelectMany(x => x.GetTypes().WhereClassesImplement(typeof(IDependencyModule)))
-            .ToArray();
+        PluginBootstrapper.LoadExternalPlugins(StartupErrors);
     }
     
     public static PhotinoBlazorApp CreateApp(string[] args)
     {
+        LoadAllPluginModules();
         var appBuilder = PhotinoBlazorAppBuilder.CreateDefault(args);
-        var dependencyModules = GetAllDependencyModules();
+        var dependencyModules = AppBootstrapper.GetAllDependencyModules();
         appBuilder.Services.AddModules(dependencyModules);
         appBuilder.RootComponents.Add<App>("#app");
         return appBuilder.Build();
     }
 
-    [STAThread]
-    public static void Main(string[] args)
+    public static void RunApp(PhotinoBlazorApp app)
     {
-        var app = CreateApp(args);
-        var pluginBootstrapper = new PluginBootstrapper(app.Services);
-        var logger = app.Services.GetService<ILogger<Program>>();
-        var eventBus = app.Services.GetService<IEventBus>();
-        
-        logger.Information("Starting Up Strem");
-        foreach (var error in StartupErrors)
-        { logger.Error(error); }
-        
-        AppDomain.CurrentDomain.UnhandledException += (sender, error) =>
-        {
-            logger.Error(error?.ExceptionObject?.ToString() ?? "Unknown Error");
-            app.MainWindow.OpenAlertWindow("Fatal exception", error?.ExceptionObject?.ToString() ?? "Unknown Error");
-        };
-
-        logger.Information("Starting Plugins");
-        Task.Run(async () => await pluginBootstrapper.StartAllPlugins(app.Services)).Wait();
-        
-        //TODO: Solve this hack with webhost service collection separation
-        WebHostHackExtensions.EventBus = eventBus;
-        WebHostHackExtensions.ServiceLocator = app.Services;
-        
-        var webHost = app.Services.GetService<IApiWebHost>();
-        logger.Information("Starting Internal Host");
-        webHost.StartHost();
-        logger.Information($"Started Internal Host: http://localhost:{ApiHostConfiguration.ApiHostPort}");
-
         var appLauncher = app.MainWindow
             .SetTitle("Strem")
             .SetSize(1920, 1080)
@@ -107,16 +58,14 @@ public class Program
         appLauncher = appLauncher.SetContextMenuEnabled(false);
 #endif
         appLauncher.Load("./wwwroot/index.html");
-        
-        logger.Information("Starting Flow Execution Engine");
-        Task.Run(async () => await pluginBootstrapper.StartExecutionEngine(app.Services)).Wait();
-        logger.Information("Started Flow Execution Engine");
-        
-        logger.Information("Strem Initialized");
-        eventBus.PublishAsync(new ApplicationStartedEvent());
+        Task.Run(() => AppBootstrapper.StartAppSetup(app, appLauncher, StartupErrors.ToArray()));
         app.Run();
-        eventBus.Publish(new ApplicationClosingEvent());
-        webHost.StopHost();
-        logger.Information("Strem Stopped");
+    }
+
+    [STAThread]
+    public static void Main(string[] args)
+    {
+        var app = CreateApp(args);
+        RunApp(app);
     }
 }
