+++ src/Strem/wwwroot/index.html
@@ -11,7 +11,7 @@
         <link href="css/app-styles.less" type="text/css" rel="stylesheet/less" />
     </head>    
     <body>
-        <div id="app">Loading...</div>
+        <div id="app" class="has-background-dark"></div>
 
         <script src="js/less.min.js" ></script>
         <script src="_framework/blazor.webview.js"></script>
